Typo fixGet rid of postfix system dependency for outgoing emails.
The idea to have to check the queue to know about errors of setup is not
working in the field when using a laptop that can often change its network
properties. This setup needs an easy way to setup a single smtp relay, and
must warn immediately when a mail is not properly sent.